Complete ComfyUI 0.35.0 integration
Adapt the merged upstream modules and tests to the package layout: LTX
keyframe, SenseNova and sparse attention nodes import through the
comfy_extras.nodes package, SenseNova's ldm and text encoder modules use
relative imports with the transformers compatibility guard, and the
relocated upstream tests resolve nodes, folder_paths and latent_preview
through the fork's modules.

Carry upstream's non-textual changes into the fork's equivalents: custom
node loaders drop display names for ignored nodes (V1 and V3), prestartup
timings are logged, ROCm builds get the aimdo OCL_SET_SVM_SIZE quota, the
DynamicVRAM warning names the recommended Torch, Porter-Duff DARKEN,
LIGHTEN, MULTIPLY and OVERLAY blend with source-over alpha, and Quantize's
Bayer path hands PIL a uint8 array again.

The Comfy Compiler flags are exposed through the Typer CLI and
Configuration; MiniMax H3 keeps its packed layout out of the pipeline
hand-off; the tensor-parallel out_proj is restored after the merge
re-added the dense one. New LTX-2.5, Pixal3D multiview, Krea-2, SCAIL-2,
Anima and MiniMax H3 turbo files join the known-model tables, SenseNova
U1.5 maps to the HiDream-O1 inference fixture until a natively loadable
checkpoint is published, and the version is 0.35.0.

+++ comfy_extras/nodes/nodes_post_processing.py
@@ -156,7 +156,7 @@ class Quantize(io.ComfyNode):
         result.add_(tiled_matrix[:result.shape[0], :result.shape[1]]).clamp_(0, 255)
         result = result.to(dtype=torch.uint8)
 
-        im = Image.fromarray(result.float().cpu().numpy())
+        im = Image.fromarray(result.cpu().numpy())
         im = im.quantize(palette=pal_im, dither=Image.Dither.NONE)
         return im
